+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 1.x branch
 ## 1.5 branch
 ### 1.5.9
+* `string_ends_with()`, `string_contains()` and `string_starts_with()` are now
+    deprecated. Use instead `str_ends_with()`, `str_contains()` and `str_starts_with()`;
 * updated for Symfony `6.0` components.
 
 ### 1.5.8